Playlist functions and some optimizations -Modify int from SongInfoContainer to byte to save a little bit of memory -Modifying SongsPage to download song instead of generate Qr -Fixing some bugs when getting NCS info -Adding Playlist logic

+++ Stereux/SongInfoContainer.xaml.cs
@@ -11,7 +11,7 @@ namespace Stereux
     /// </summary>
     public partial class SongInfoContainer : Page
     {
-        private int _infoIndex;
+        private byte _infoIndex;
 
         // For each array, the order is 1. Stereux, 2. Source, 3. Song
         private string[] _titles = { "Download Stereux", "Music provided by:", "Get more info" };
